# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: OneLife Finance API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /loans:
    get:
      operationId: listLoans
      tags: [loans]
      summary: List loans
      responses:
        "200":
          description: List of loans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Loan"
    post:
      operationId: createLoan
      tags: [loans]
      summary: Create a loan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanInput"
      responses:
        "201":
          description: Created loan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /loans/{id}:
    get:
      operationId: getLoan
      tags: [loans]
      summary: Get a loan with its EMI schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Loan detail with EMIs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanDetail"
        "404":
          description: Not found
    patch:
      operationId: updateLoan
      tags: [loans]
      summary: Update a loan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanUpdate"
      responses:
        "200":
          description: Updated loan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
        "404":
          description: Not found
    delete:
      operationId: deleteLoan
      tags: [loans]
      summary: Delete a loan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /insurances:
    get:
      operationId: listInsurances
      tags: [insurances]
      summary: List insurance policies
      responses:
        "200":
          description: List of insurance policies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Insurance"
    post:
      operationId: createInsurance
      tags: [insurances]
      summary: Create an insurance policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InsuranceInput"
      responses:
        "201":
          description: Created insurance policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Insurance"
  /insurances/{id}:
    patch:
      operationId: updateInsurance
      tags: [insurances]
      summary: Update an insurance policy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InsuranceUpdate"
      responses:
        "200":
          description: Updated insurance policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Insurance"
        "404":
          description: Not found
    delete:
      operationId: deleteInsurance
      tags: [insurances]
      summary: Delete an insurance policy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /investments:
    get:
      operationId: listInvestments
      tags: [investments]
      summary: List investments
      responses:
        "200":
          description: List of investments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Investment"
    post:
      operationId: createInvestment
      tags: [investments]
      summary: Create an investment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvestmentInput"
      responses:
        "201":
          description: Created investment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investment"
  /investments/{id}:
    patch:
      operationId: updateInvestment
      tags: [investments]
      summary: Update an investment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvestmentUpdate"
      responses:
        "200":
          description: Updated investment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Investment"
        "404":
          description: Not found
    delete:
      operationId: deleteInvestment
      tags: [investments]
      summary: Delete an investment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List savings goals
      responses:
        "200":
          description: List of goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a savings goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a savings goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "404":
          description: Not found
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a savings goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  schemas:
...
    EmiStatus:
      type: string
      enum: [paid, pending, overdue, partial]
    LoanType:
      type: string
      enum: [home, car, personal, gold, education, other]
    LoanStatus:
      type: string
      enum: [active, closed]
    InsuranceType:
      type: string
      enum: [health, life, car, house, term]
    InsuranceStatus:
      type: string
      enum: [active, expired]
    InvestmentType:
      type: string
      enum: [mutual_fund, stocks, ppf, nps, fd, gold, crypto]
    GoalType:
      type: string
      enum: [emergency_fund, vacation, car, house, kids_education, retirement, other]
...
    Loan:
      type: object
      required:
        [id, name, loanType, principalAmount, outstandingAmount, interestRate,
         emiAmount, tenureMonths, monthsRemaining, startDate, status, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        loanType:
          $ref: "#/components/schemas/LoanType"
        principalAmount: { type: number }
        outstandingAmount: { type: number }
        interestRate: { type: number }
        emiAmount: { type: number }
        tenureMonths: { type: integer }
        monthsRemaining: { type: integer }
        startDate: { type: string, format: date }
        status:
          $ref: "#/components/schemas/LoanStatus"
        bankName: { type: ["string", "null"] }
        bankLogoUrl: { type: ["string", "null"] }
        disbursementDocUrl: { type: ["string", "null"] }
        repaymentScheduleDocUrl: { type: ["string", "null"] }
        penaltyRate: { type: number }
        createdAt: { type: string, format: date-time }
    LoanDetail:
      allOf:
        - $ref: "#/components/schemas/Loan"
        - type: object
          required: [emis, principalPaid, interestPaid]
          properties:
            principalPaid: { type: number }
            interestPaid: { type: number }
            emis:
              type: array
              items:
                $ref: "#/components/schemas/Emi"
    LoanInput:
      type: object
      required: [name, loanType, principalAmount, interestRate, emiAmount, tenureMonths, startDate]
      properties:
        name: { type: string, minLength: 1 }
        loanType:
          $ref: "#/components/schemas/LoanType"
        principalAmount: { type: number, minimum: 0 }
        outstandingAmount: { type: number, minimum: 0 }
        interestRate: { type: number, minimum: 0 }
        emiAmount: { type: number, minimum: 0 }
        tenureMonths: { type: integer, minimum: 1 }
        startDate: { type: string, format: date }
        bankName: { type: string }
        bankLogoUrl: { type: string }
        disbursementDocUrl: { type: string }
        repaymentScheduleDocUrl: { type: string }
        penaltyRate: { type: number, minimum: 0, default: 2.0 }
    LoanUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        loanType:
          $ref: "#/components/schemas/LoanType"
        principalAmount: { type: number, minimum: 0 }
        outstandingAmount: { type: number, minimum: 0 }
        interestRate: { type: number, minimum: 0 }
        emiAmount: { type: number, minimum: 0 }
        tenureMonths: { type: integer, minimum: 1 }
        startDate: { type: string, format: date }
        bankName: { type: ["string", "null"] }
        bankLogoUrl: { type: ["string", "null"] }
        disbursementDocUrl: { type: ["string", "null"] }
        repaymentScheduleDocUrl: { type: ["string", "null"] }
        penaltyRate: { type: number, minimum: 0 }
        status:
          $ref: "#/components/schemas/LoanStatus"
    Emi:
      type: object
      required: [id, loanId, loanName, dueDate, paidDate, amount, status]
      properties:
        id: { type: integer }
        loanId: { type: integer }
        loanName: { type: string }
        dueDate: { type: string, format: date }
        paidDate: { type: ["string", "null"], format: date }
        amount: { type: number }
        status:
          $ref: "#/components/schemas/EmiStatus"
        penaltyAmount: { type: number }
        overdueDays: { type: integer }
...
    Insurance:
      type: object
      required:
        [id, name, insuranceType, provider, premiumAmount, coverageAmount,
         renewalDate, status, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        insuranceType:
          $ref: "#/components/schemas/InsuranceType"
        provider: { type: string }
        premiumAmount: { type: number }
        coverageAmount: { type: number }
        renewalDate: { type: string, format: date }
        policyNumber: { type: ["string", "null"] }
        status:
          $ref: "#/components/schemas/InsuranceStatus"
        createdAt: { type: string, format: date-time }
    InsuranceInput:
      type: object
      required: [name, insuranceType, provider, premiumAmount, coverageAmount, renewalDate]
      properties:
        name: { type: string, minLength: 1 }
        insuranceType:
          $ref: "#/components/schemas/InsuranceType"
        provider: { type: string, minLength: 1 }
        premiumAmount: { type: number, minimum: 0 }
        coverageAmount: { type: number, minimum: 0 }
        renewalDate: { type: string, format: date }
        policyNumber: { type: string }
    InsuranceUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        insuranceType:
          $ref: "#/components/schemas/InsuranceType"
        provider: { type: string, minLength: 1 }
        premiumAmount: { type: number, minimum: 0 }
        coverageAmount: { type: number, minimum: 0 }
        renewalDate: { type: string, format: date }
        policyNumber: { type: ["string", "null"] }
        status:
          $ref: "#/components/schemas/InsuranceStatus"
    Investment:
      type: object
      required:
        [id, name, investmentType, investedAmount, currentValue, purchaseDate, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        investmentType:
          $ref: "#/components/schemas/InvestmentType"
        investedAmount: { type: number }
        currentValue: { type: number }
        purchaseDate: { type: string, format: date }
        xirr: { type: ["number", "null"] }
        createdAt: { type: string, format: date-time }
    InvestmentInput:
      type: object
      required: [name, investmentType, investedAmount, currentValue, purchaseDate]
      properties:
        name: { type: string, minLength: 1 }
        investmentType:
          $ref: "#/components/schemas/InvestmentType"
        investedAmount: { type: number, minimum: 0 }
        currentValue: { type: number, minimum: 0 }
        purchaseDate: { type: string, format: date }
        xirr: { type: number }
    InvestmentUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        investmentType:
          $ref: "#/components/schemas/InvestmentType"
        investedAmount: { type: number, minimum: 0 }
        currentValue: { type: number, minimum: 0 }
        purchaseDate: { type: string, format: date }
        xirr: { type: ["number", "null"] }
    Goal:
      type: object
      required: [id, name, goalType, targetAmount, currentAmount, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        goalType:
          $ref: "#/components/schemas/GoalType"
        targetAmount: { type: number }
        currentAmount: { type: number }
        targetDate: { type: ["string", "null"], format: date }
        createdAt: { type: string, format: date-time }
    GoalInput:
      type: object
      required: [name, goalType, targetAmount]
      properties:
        name: { type: string, minLength: 1 }
        goalType:
          $ref: "#/components/schemas/GoalType"
        targetAmount: { type: number, minimum: 0 }
        currentAmount: { type: number, minimum: 0, default: 0 }
        targetDate: { type: string, format: date }
    GoalUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        goalType:
          $ref: "#/components/schemas/GoalType"
        targetAmount: { type: number, minimum: 0 }
        currentAmount: { type: number, minimum: 0 }
        targetDate: { type: ["string", "null"], format: date }
...