# lib/api-zod/src/generated/types/insuranceStatus.js
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * OneLife Finance API specification
 * OpenAPI spec version: 0.1.0
 */
export const InsuranceStatus = {
    active: 'active',
    expired: 'expired',
};